backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.4.2
setuptools
wheel 